+++ src/Client/ui/components/card.jsx
@@ -1,30 +1,39 @@
 import React from "react";
+import PropTypes from 'prop-types';
+import CardArt from "./card-art";
 import { orNothing } from "../../../utils/help";
-import { renderCard } from "../../../utils/render";
 import "../styles/card.css";
 
-const Card = (props) => {
+const Card = ({card, highlight, extraClass, click, children}) => {
 
     const selectCard = () => {
-        if(props.highlight && props.card){
-            props.highlight(props.card)
+        if(highlight && card){
+            highlight(card)
         }
     }
 
     const checkAndRender = () => {
-        if(props.card){
-            let reveal = typeof props.card.flipped !== 'undefined' ? !props.card.flipped : false;
-            return renderCard(props.card, reveal)
+        if(card){
+            let reveal = typeof card.flipped !== 'undefined' ? !card.flipped : false;
+            return <CardArt card={card} reveal={reveal}/>
         }
     }
 
     return(
-    <div className={`hoverable card-holder `+orNothing(props.extraClass)} 
-    onClick={props.click}
+    <div className={`hoverable card-holder `+orNothing(extraClass)} 
+    onClick={click}
     onMouseEnter={(e) => selectCard()}>
         {checkAndRender()}
-        {props.children}
+        {children}
     </div>
 )}
+
+Card.propTypes = {
+    card: PropTypes.object, 
+    highlight: PropTypes.func, 
+    extraClass: PropTypes.string, 
+    click: PropTypes.func, 
+    children: PropTypes.node
+}
 
 export default Card;