+++ src/Client/ui/components/hand.jsx
@@ -1,18 +1,19 @@
 import React from "react";
+import PropTypes from 'prop-types';
 import Card from "./card";
 import { Origin } from "../../../models/enums";
 import "../styles/hand.css";
 
-const Hand = (props) => {
+const Hand = ({list, selected, menuClick, highlight}) => {
     const handMenu = (e, i) => {
-        if(props.menuClick){
-            props.menuClick(e, i);
+        if(menuClick){
+            menuClick(e, i);
         }
     }
 
     const isSelected = (index) => {
-        if(props.selected){
-            let thisSelection = props.selected.origin[Origin.HAND]
+        if(selected){
+            let thisSelection = selected.origin[Origin.HAND]
             if(thisSelection === index){
                 return " selected";
             }
@@ -22,13 +23,20 @@ const Hand = (props) => {
 
     return(
     <div className="hand">
-        {props.list.map((card, index) => (
+        {list.map((card, index) => (
         <Card key={index}
         card={card}
-        highlight={props.highlight}
+        highlight={highlight}
         extraClass={isSelected(index)} 
         click={(e) => handMenu(e, index)}/>
     ))}</div>
 )}
+
+Hand.propTypes = {
+    list: PropTypes.arrayOf(PropTypes.object), 
+    selected: PropTypes.object, 
+    menuClick: PropTypes.func, 
+    highlight: PropTypes.func
+}
 
 export default Hand;