+++ src/Client/ui/components/jar.jsx
@@ -1,13 +1,20 @@
 import React from "react";
+import PropTypes from 'prop-types';
 import { orNothing } from "../../../utils/help";
 import "../styles/jar.css";
 
-const Jar = (props) => (
+const Jar = ({click, extraClass, children}) => (
     <div 
-    className={"hoverable jar "+orNothing(props.extraClass)} 
-    onClick={props.click}>
-        {props.children}
+    className={"hoverable jar "+orNothing(extraClass)} 
+    onClick={click}>
+        {children}
         </div>
 )
+
+Jar.propTypes = {
+    click: PropTypes.func, 
+    extraClass: PropTypes.string, 
+    children: PropTypes.node
+}
 
 export default Jar;